Update mash with new sharedmem

+++ PopPUNK/refine.py
@@ -83,7 +83,7 @@ def refineFit(distMat, sample_names, start_s, mean0, mean1,
     # Move distMat into shared memory
     with SharedMemoryManager() as smm:
         shm_distMat = smm.SharedMemory(create = True, size = distMat.nbytes, name = 'shm_distMat')
-        distances_shared = np.ndarray(distMat.shape, dtype = np.float32, buffer = shm_distMat.buf)
+        distances_shared = np.ndarray(distMat.shape, dtype = distMat.dtype, buffer = shm_distMat.buf)
         distances_shared[:] = distMat[:]
         
         with Pool(processes = num_processes) as pool:
